Implement a button to disable tracks
This commit adds a column to the MusicManager window with a button to
disable individual tracks from playing, which is accomplished by setting
the SongDef commonality to 0. Re-enabling the track sets commonality to
1 (the default). In keeping with existing convention, clicking the
column header will sort tracks by disabled status, and shift-clicking
the header will mass disable/enable tracks.